remove the legacy way of creating adapters
