+++ templates/index.ts
@@ -68,12 +68,10 @@ const templates: { name: string, templateFunction: TemplateFunction }[] = [
 	{ name: "io-package.json.ts", templateFunction: require("./io-package.json") },
 	{ name: "lib/adapter-config.dts.ts", templateFunction: require("./lib/adapter-config.dts") },
 	{ name: "LICENSE.ts", templateFunction: require("./LICENSE") },
-	{ name: "main.es6.js.ts", templateFunction: require("./main.es6.js") },
 	{ name: "main.js.ts", templateFunction: require("./main.js") },
 	{ name: "main.test.js.ts", templateFunction: require("./main.test.js") },
 	{ name: "package.json.ts", templateFunction: require("./package.json") },
 	{ name: "README.md.ts", templateFunction: require("./README.md") },
-	{ name: "src/main.es6.ts", templateFunction: require("./src/main.es6") },
 	{ name: "src/main.test.ts.ts", templateFunction: require("./src/main.test.ts") },
 	{ name: "src/main.ts.ts", templateFunction: require("./src/main.ts") },
 	{ name: "test/_eslintrc.json.ts", templateFunction: require("./test/_eslintrc.json") },
