+++ templates/main.js.ts
@@ -1,12 +1,13 @@
 import { AdapterSettings, getDefaultAnswer } from "../src/lib/core/questions";
 import type { TemplateFunction } from "../src/lib/createAdapter";
+import { kebabCaseToUpperCamelCase } from "../src/lib/tools";
 
-export = (async answers => {
+const templateFunction: TemplateFunction = async answers => {
 
 	const useJavaScript = answers.language === "JavaScript";
-	const useES6Class = answers.es6class === "yes";
-	if (!useJavaScript || useES6Class) return;
+	if (!useJavaScript) return;
 
+	const className = kebabCaseToUpperCamelCase(answers.adapterName);
 	const adapterSettings: AdapterSettings[] = answers.adapterSettings || getDefaultAnswer("adapterSettings")!;
 	const quote = answers.quotes === "double" ? '"' : "'";
 	const t = answers.indentation === "Space (4)" ? "    " : "\t";
@@ -25,148 +26,166 @@ const utils = require("@iobroker/adapter-core");
 // Load your modules here, e.g.:
 // const fs = require("fs");
 
-/**
- * The adapter instance
- * @type {ioBroker.Adapter}
- */
-let adapter;
+class ${className} extends utils.Adapter {
 
-/**
- * Starts the adapter instance
- * @param {Partial<utils.AdapterOptions>} [options]
- */
-function startAdapter(options) {
-	// Create the adapter and define its methods
-	return adapter = utils.adapter(Object.assign({}, options, {
-		name: "${answers.adapterName}",
+	/**
+	 * @param {Partial<utils.AdapterOptions>} [options={}]
+	 */
+	constructor(options) {
+		super({
+			...options,
+			name: "${answers.adapterName}",
+		});
+		this.on("ready", this.onReady.bind(this));
+		this.on("stateChange", this.onStateChange.bind(this));
+		// this.on(${quote}objectChange${quote}, this.onObjectChange.bind(this));
+		// this.on(${quote}message${quote}, this.onMessage.bind(this));
+		this.on("unload", this.onUnload.bind(this));
+	}
 
-		// The ready callback is called when databases are connected and adapter received configuration.
-		// start here!
-		ready: main, // Main method defined below for readability
-
-		// is called when adapter shuts down - callback has to be called under any circumstances!
-		unload: (callback) => {
-			try {
-				// Here you must clear all timeouts or intervals that may still be active
-				// clearTimeout(timeout1);
-				// clearTimeout(timeout2);
-				// ...
-				// clearInterval(interval1);
-
-				callback();
-			} catch (e) {
-				callback();
-			}
-		},
-
-		// If you need to react to object changes, uncomment the following method.
-		// You also need to subscribe to the objects with \`adapter.subscribeObjects\`, similar to \`adapter.subscribeStates\`.
-		// objectChange: (id, obj) => {
-		// ${t}if (obj) {
-		// ${t}${t}// The object was changed
-		// ${t}${t}adapter.log.info(\`object \$\{id} changed: \$\{JSON.stringify(obj)}\`);
-		// ${t}} else {
-		// ${t}${t}// The object was deleted
-		// ${t}${t}adapter.log.info(\`object \$\{id} deleted\`);
-		// ${t}}
-		// },
-
-		// is called if a subscribed state changes
-		stateChange: (id, state) => {
-			if (state) {
-				// The state was changed
-				adapter.log.info(\`state \$\{id} changed: \$\{state.val} (ack = \$\{state.ack})\`);
-			} else {
-				// The state was deleted
-				adapter.log.info(\`state \$\{id} deleted\`);
-			}
-		},
-
-		// If you need to accept messages in your adapter, uncomment the following block.
-		// /**
-		//  * Some message was sent to this instance over message box. Used by email, pushover, text2speech, ...
-		//  * Using this method requires "common.messagebox" property to be set to true in io-package.json
-		//  */
-		// message: (obj) => {
-		// ${t}if (typeof obj === ${quote}object${quote} && obj.message) {
-		// ${t}${t}if (obj.command === ${quote}send${quote}) {
-		// ${t}${t}${t}// e.g. send email or pushover or whatever
-		// ${t}${t}${t}adapter.log.info(${quote}send command${quote});
-
-		// ${t}${t}${t}// Send response in callback if required
-		// ${t}${t}${t}if (obj.callback) adapter.sendTo(obj.from, obj.command, ${quote}Message received${quote}, obj.callback);
-		// ${t}${t}}
-		// ${t}}
-		// },
-	}));
-}
-
-async function main() {
+	/**
+	 * Is called when databases are connected and adapter received configuration.
+	 */
+	async onReady() {
+		// Initialize your adapter here
 
 ${answers.connectionIndicator === "yes" ? `
-	// Reset the connection indicator during startup
-	await adapter.setStateAsync("info.connection", false, true);
+		// Reset the connection indicator during startup
+		this.setState("info.connection", false, true);
 ` : ""}
 
-	// The adapters config (in the instance object everything under the attribute "native") is accessible via
-	// adapter.config:
-${adapterSettings.map(s => `\tadapter.log.info("config ${s.key}: " + adapter.config.${s.key});`).join("\n")}
+		// The adapters config (in the instance object everything under the attribute "native") is accessible via
+		// this.config:
+${adapterSettings.map(s => `\t\tthis.log.info("config ${s.key}: " + this.config.${s.key});`).join("\n")}
 
-	/*
+		/*
 		For every state in the system there has to be also an object of type state
 		Here a simple template for a boolean variable named "testVariable"
 		Because every adapter instance uses its own unique namespace variable names can't collide with other adapters variables
-	*/
-	await adapter.setObjectNotExistsAsync("testVariable", {
-		type: "state",
-		common: {
-			name: "testVariable",
-			type: "boolean",
-			role: "indicator",
-			read: true,
-			write: true,
-		},
-		native: {},
-	});
+		*/
+		await this.setObjectNotExistsAsync("testVariable", {
+			type: "state",
+			common: {
+				name: "testVariable",
+				type: "boolean",
+				role: "indicator",
+				read: true,
+				write: true,
+			},
+			native: {},
+		});
 
-	// In order to get state updates, you need to subscribe to them. The following line adds a subscription for our variable we have created above.
-	adapter.subscribeStates("testVariable");
-	// You can also add a subscription for multiple states. The following line watches all states starting with "lights."
-	// adapter.subscribeStates(${quote}lights.*${quote});
-	// Or, if you really must, you can also watch all states. Don't do this if you don't need to. Otherwise this will cause a lot of unnecessary load on the system:
-	// adapter.subscribeStates(${quote}*${quote});
+		// In order to get state updates, you need to subscribe to them. The following line adds a subscription for our variable we have created above.
+		this.subscribeStates("testVariable");
+		// You can also add a subscription for multiple states. The following line watches all states starting with "lights."
+		// this.subscribeStates(${quote}lights.*${quote});
+		// Or, if you really must, you can also watch all states. Don't do this if you don't need to. Otherwise this will cause a lot of unnecessary load on the system:
+		// this.subscribeStates(${quote}*${quote});
 
-	/*
-		setState examples
-		you will notice that each setState will cause the stateChange event to fire (because of above subscribeStates cmd)
-	*/
-	// the variable testVariable is set to true as command (ack=false)
-	await adapter.setStateAsync("testVariable", true);
+		/*
+			setState examples
+			you will notice that each setState will cause the stateChange event to fire (because of above subscribeStates cmd)
+		*/
+		// the variable testVariable is set to true as command (ack=false)
+		await this.setStateAsync("testVariable", true);
 
-	// same thing, but the value is flagged "ack"
-	// ack should be always set to true if the value is received from or acknowledged from the target system
-	await adapter.setStateAsync("testVariable", { val: true, ack: true });
+		// same thing, but the value is flagged "ack"
+		// ack should be always set to true if the value is received from or acknowledged from the target system
+		await this.setStateAsync("testVariable", { val: true, ack: true });
 
-	// same thing, but the state is deleted after 30s (getState will return null afterwards)
-	await adapter.setStateAsync("testVariable", { val: true, ack: true, expire: 30 });
+		// same thing, but the state is deleted after 30s (getState will return null afterwards)
+		await this.setStateAsync("testVariable", { val: true, ack: true, expire: 30 });
 
-	// examples for the checkPassword/checkGroup functions
-	adapter.checkPassword("admin", "iobroker", (res) => {
-		adapter.log.info("check user admin pw iobroker: " + res);
-	});
+		// examples for the checkPassword/checkGroup functions
+		let result = await this.checkPasswordAsync("admin", "iobroker");
+		this.log.info("check user admin pw iobroker: " + result);
 
-	adapter.checkGroup("admin", "admin", (res) => {
-		adapter.log.info("check group user admin group admin: " + res);
-	});
+		result = await this.checkGroupAsync("admin", "admin");
+		this.log.info("check group user admin group admin: " + result);
+	}
+
+	/**
+	 * Is called when adapter shuts down - callback has to be called under any circumstances!
+	 * @param {() => void} callback
+	 */
+	onUnload(callback) {
+		try {
+			// Here you must clear all timeouts or intervals that may still be active
+			// clearTimeout(timeout1);
+			// clearTimeout(timeout2);
+			// ...
+			// clearInterval(interval1);
+
+			callback();
+		} catch (e) {
+			callback();
+		}
+	}
+
+	// If you need to react to object changes, uncomment the following block and the corresponding line in the constructor.
+	// You also need to subscribe to the objects with \`this.subscribeObjects\`, similar to \`this.subscribeStates\`.
+	// /**
+	//  * Is called if a subscribed object changes
+	//  * @param {string} id
+	//  * @param {ioBroker.Object | null | undefined} obj
+	//  */
+	// onObjectChange(id, obj) {
+	// ${t}if (obj) {
+	// ${t}${t}// The object was changed
+	// ${t}${t}this.log.info(\`object \${id} changed: \${JSON.stringify(obj)}\`);
+	// ${t}} else {
+	// ${t}${t}// The object was deleted
+	// ${t}${t}this.log.info(\`object \${id} deleted\`);
+	// ${t}}
+	// }
+
+	/**
+	 * Is called if a subscribed state changes
+	 * @param {string} id
+	 * @param {ioBroker.State | null | undefined} state
+	 */
+	onStateChange(id, state) {
+		if (state) {
+			// The state was changed
+			this.log.info(\`state \${id} changed: \${state.val} (ack = \${state.ack})\`);
+		} else {
+			// The state was deleted
+			this.log.info(\`state \${id} deleted\`);
+		}
+	}
+
+	// If you need to accept messages in your adapter, uncomment the following block and the corresponding line in the constructor.
+	// /**
+	//  * Some message was sent to this instance over message box. Used by email, pushover, text2speech, ...
+	//  * Using this method requires "common.messagebox" property to be set to true in io-package.json
+	//  * @param {ioBroker.Message} obj
+	//  */
+	// onMessage(obj) {
+	// ${t}if (typeof obj === ${quote}object${quote} && obj.message) {
+	// ${t}${t}if (obj.command === ${quote}send${quote}) {
+	// ${t}${t}${t}// e.g. send email or pushover or whatever
+	// ${t}${t}${t}this.log.info(${quote}send command${quote});
+
+	// ${t}${t}${t}// Send response in callback if required
+	// ${t}${t}${t}if (obj.callback) this.sendTo(obj.from, obj.command, ${quote}Message received${quote}, obj.callback);
+	// ${t}${t}}
+	// ${t}}
+	// }
+
 }
 
 if (require.main !== module) {
-	// Export startAdapter in compact mode
-	module.exports = startAdapter;
+	// Export the constructor in compact mode
+	/**
+	 * @param {Partial<utils.AdapterOptions>} [options={}]
+	 */
+	module.exports = (options) => new ${className}(options);
 } else {
 	// otherwise start the instance directly
-	startAdapter();
+	new ${className}();
 }
 `;
 	return template.trim();
-}) as TemplateFunction;
+};
+templateFunction.customPath = "main.js";
+export = templateFunction;
